+++ test.f.cjs
@@ -22,7 +22,8 @@ require('./commonjs/build/test.f.cjs')
 require('./types/range/test.f.cjs')
 require('./html/test.f.cjs')
 require('./text/test.f.cjs')
-require('./text/encoding/test.f.cjs')
+require('./text/encoding/utf8/test.f.cjs')
+require('./text/encoding/utf16/test.f.cjs')
 require('./com/cs/test.f.cjs')
 require('./com/cpp/test.f.cjs')
 require('./nodejs/version/test.f.cjs')
