+++ text/encoding/utf8/module.f.cjs
@@ -1,33 +1,17 @@
-const list = require('../../types/list/module.f.cjs')
-const operator = require('../../types/function/operator/module.f.cjs')
-const array = require('../../types/array/module.f.cjs')
-const { contains } = require('../../types/range/module.f.cjs')
-const { compose } = require('../../types/function/module.f.cjs')
-const { map, flat, stateScan, concat, reduce, toArray, flatMap } = list
+const list = require('../../../types/list/module.f.cjs')
+const operator = require('../../../types/function/operator/module.f.cjs')
+const array = require('../../../types/array/module.f.cjs')
+const { flatMap } = list
 
 /** @typedef {u8|undefined} ByteOrEof */
-
-/** @typedef {u16|undefined} WordOrEof */
 
 /** @typedef {array.Array1<number>|array.Array2<number>|array.Array3<number>} Utf8NonEmptyState */
 
 /** @typedef {undefined|Utf8NonEmptyState} Utf8State */
 
-/** @typedef {undefined|number} Utf16State */
-
 /** @typedef {number} u8 */
 
-/** @typedef {number} u16 */
-
 /** @typedef {number} i32 */
-
-/** @type {(a:number) => boolean} */
-const isBmpCodePoint = a => a >= 0x0000 && a <= 0xd7ff || a >= 0xe000 && a <= 0xffff
-
-const isHighSurrogate = contains([0xd800, 0xdbff])
-
-/** @type {(a:number) => boolean} */
-const isLowSurrogate = contains([0xdc00, 0xdfff])
 
 const errorMask = 0b1000_0000_0000_0000_0000_0000_0000_0000
 
@@ -47,21 +31,7 @@ const codePointToUtf8 = input =>
     return [errorMask]
 }
 
-/** @type {(input:i32) => list.List<u16>} */
-const codePointToUtf16 = input =>
-{
-    if (isBmpCodePoint(input)) { return [input] }
-    if (input >= 0x010000 && input <= 0x10ffff) {
-        const high = ((input - 0x10000) >> 10) + 0xd800
-        const low = ((input - 0x10000) & 0b0011_1111_1111) + 0xdc00
-        return [high, low]
-    }
-    return [input & 0xffff]
-}
-
-const codePointListToUtf8List = flatMap(codePointToUtf8)
-
-const codePointListToUtf16List = flatMap(codePointToUtf16)
+const fromCodePointList = flatMap(codePointToUtf8)
 
 /** @type {(state: Utf8NonEmptyState) => i32}*/
 const utf8StateToError = state => {
@@ -116,60 +86,11 @@ const utf8EofToCodePointOp = state => {
 const utf8ByteOrEofToCodePointOp = state => input => input === undefined ? utf8EofToCodePointOp(state) : utf8ByteToCodePointOp(state)(input)
 
 /** @type {(input: list.List<u8>) => list.List<i32>} */
-const utf8ListToCodePointList = input => list.flat(list.stateScan(utf8ByteOrEofToCodePointOp)(undefined)(list.concat(/** @type {list.List<ByteOrEof>} */(input))([undefined])))
-
-/** @type {operator.StateScan<u16, Utf16State, list.List<i32>>} */
-const utf16ByteToCodePointOp = state => byte => {
-    if (byte < 0x00 || byte > 0xffff) {
-        return [[0xffffffff], state]
-    }
-    if (state === undefined) {
-        if (isBmpCodePoint(byte)) { return [[byte], undefined] }
-        if (isHighSurrogate(byte)) { return [[], byte] }
-        return [[byte | errorMask], undefined]
-    }
-    if (isLowSurrogate(byte)) {
-        const high = state - 0xd800
-        const low = byte - 0xdc00
-        return [[(high << 10) + low + 0x10000], undefined]
-    }
-    if (isBmpCodePoint(byte)) { return [[state | errorMask, byte], undefined] }
-    if (isHighSurrogate(byte)) { return [[state | errorMask], byte] }
-    return [[state | errorMask, byte | errorMask], undefined]
-}
-
-/** @type {(state: Utf16State) => readonly[list.List<i32>, Utf16State]} */
-const utf16EofToCodePointOp = state => [state === undefined ? undefined : [state | errorMask],  undefined]
-
-/** @type {operator.StateScan<WordOrEof, Utf16State, list.List<i32>>} */
-const utf16ByteOrEofToCodePointOp = state => input => input === undefined ? utf16EofToCodePointOp(state) : utf16ByteToCodePointOp(state)(input)
-
-/** @type {(input: list.List<u16>) => list.List<i32>} */
-const utf16ListToCodePointList = input => flat(stateScan(utf16ByteOrEofToCodePointOp)(undefined)(concat(/** @type {list.List<WordOrEof>} */(input))([undefined])))
-
-/** @type {(s: string) => list.List<u16>} */
-const stringToUtf16List = s => {
-    /** @type {(i: number) => list.Result<number>} */
-    const at = i => {
-        const first = s.charCodeAt(i)
-        return isNaN(first) ? undefined : { first, tail: () => at(i + 1) }
-    }
-    return at(0)
-}
-
-const utf16ListToString = compose(map(String.fromCharCode))(reduce(operator.concat)(''))
+const toCodePointList = input => list.flat(list.stateScan(utf8ByteOrEofToCodePointOp)(undefined)(list.concat(/** @type {list.List<ByteOrEof>} */(input))([undefined])))
 
 module.exports = {
     /** @readonly */
-    codePointListToUtf8List,
+    fromCodePointList,
     /** @readonly */
-    codePointListToUtf16List,
-    /** @readonly */
-    utf8ListToCodePointList,
-    /** @readonly */
-    utf16ListToCodePointList,
-    /** @readonly */
-    stringToUtf16List,
-    /** @readonly */
-    utf16ListToString
+    toCodePointList
 }
